fix issue with findUp and git commit message
I found a few errors while trying to use this package, the first was:
```
The "path" argument must be one of type string, Buffer, or URL. Received type undefined
```
This was because findUp was not choosing to locate the .git directory because the type was not specified, so I added Promise.race to get either a file or directory... I think this should work!

Then the next issue was that the message did contain the JIRA issue number because the pager I use shows the branch name.